requirements.txt
# Visão Computacional
opencv-python==4.9.0.80
deepface==0.0.93
mediapipe==0.10.14
ultralytics==8.2.0          # YOLOv8
face-recognition==1.3.0

# Áudio
moviepy==1.0.3
SpeechRecognition==3.10.4
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.2             # Análise prosódia / tom de voz

# Azure Cognitive Services
azure-cognitiveservices-speech==1.38.0
azure-ai-vision-imageanalysis==1.0.0
azure-ai-language-conversations==1.1.0

# OpenAI
openai==1.30.0

# Processamento de texto / NLP
transformers==4.41.0
torch==2.3.0
nltk==3.8.1

# Dados e utilidades
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.3.0
tqdm==4.66.4

# Notebooks
jupyter==1.0.0
ipykernel==6.29.4
matplotlib==3.9.0
seaborn==0.13.2
